working test computation of main eigen vector with img2eig

+++ kernels/matrix_norm.cl
@@ -21,22 +21,22 @@ matrix_normalize(__global float* R, __global float* A, int m, int n)
     // thread variables
     float mu = 0.0f; // average of the column values
     float n_var = 0.0f; // total square distance to the average
-    
+
     // Compute the column average;
     for (int k = 0; k < m; ++k) {
         // Calculation of the column average
         mu += A[k * n + ty];
-     }
+    }
 
-     mu = mu / m;
+    mu = mu / (float)m;
 
-     // Compute the column variance
-     for (int k = 0; k < m; ++k) {
-         n_var += pown( A[k * n + ty] - mu, 2 );
-     }
+    // Compute the column variance
+    for (int k = 0; k < m; ++k) {
+        n_var += pown( A[k * n + ty] - mu, 2 );
+    }
 
-     // Normalize the matrix
-     for (int k = 0; k < m; ++k) {
+    // Normalize the matrix
+    for (int k = 0; k < m; ++k) {
         R[k * n + ty] = (A[k * n + ty] - mu) / sqrt(n_var);
-     }
+    }
 }
